requirements.txt
# 优化后的依赖列表 - 轻量精简
# 从8个减少到5个，移除未使用的依赖

# RSS 解析
feedparser>=6.0.11,<7.0

# AI API - 升级到2.x版本
openai>=2.0.0,<3.0

# 日期处理
python-dateutil>=2.8.2,<3.0

# YAML 解析
PyYAML>=6.0.1,<7.0

# 重试机制
tenacity>=8.2.0,<9.0
